+++ dataprovider/data_provider.go
@@ -1,0 +1,39 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package dataprovider
+
+import (
+	"context"
+	"github.com/elastic/beats/v7/libbeat/beat"
+	"github.com/elastic/cloudbeat/resources/fetching"
+	"github.com/elastic/cloudbeat/version"
+)
+
+type CommonDataProvider interface {
+	FetchCommonData(ctx context.Context) (CommonData, error)
+}
+
+type CommonData interface {
+	GetResourceId(fetching.ResourceMetadata) string
+	GetVersionInfo() version.CloudbeatVersionInfo
+	EnrichEvent(beat.Event) error
+}
+
+type EnvironmentCommonDataProvider interface {
+	FetchData(context.Context) (CommonData, error)
+}
